requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
